Move type definitions away from game-data.ts
