+++ src/@types/arena-tags.ts
@@ -1,7 +1,7 @@
 import type { ArenaTagTypeMap } from "#data/arena-tag";
 import type { ArenaTagType } from "#enums/arena-tag-type";
 // biome-ignore lint/correctness/noUnusedImports: TSDocs
-import type { SessionSaveData } from "#system/game-data";
+import type { SessionSaveData } from "#types/data-types";
 
 /** Subset of {@linkcode ArenaTagType}s that apply some negative effect to pokemon that switch in ({@link https://bulbapedia.bulbagarden.net/wiki/List_of_moves_that_cause_entry_hazards#List_of_traps | entry hazards} and Imprison. */
 export type EntryHazardTagType =
