+++ src/data/egg-hatch-data.ts
@@ -1,6 +1,6 @@
 import { globalScene } from "#app/global-scene";
 import type { PlayerPokemon } from "#field/pokemon";
-import type { StarterDataEntry } from "#system/game-data";
+import type { StarterDataEntry } from "#types/data-types";
 import type { DexEntry } from "#types/dex-data";
 
 /**
