+++ src/phases/title-phase.ts
@@ -14,8 +14,8 @@ import { Unlockables } from "#enums/unlockables";
 import { getBiomeKey } from "#field/arena";
 import type { Modifier } from "#modifiers/modifier";
 import { getDailyRunStarterModifiers, regenerateModifierPoolThresholds } from "#modifiers/modifier-type";
-import type { SessionSaveData } from "#system/game-data";
 import { vouchers } from "#system/voucher";
+import type { SessionSaveData } from "#types/data-types";
 import type { OptionSelectConfig, OptionSelectItem } from "#ui/abstract-option-select-ui-handler";
 import { SaveSlotUiMode } from "#ui/save-slot-select-ui-handler";
 import { isLocal, isLocalServerConnected, isNullOrUndefined } from "#utils/common";
