+++ src/system/version-migration/version-converter.ts
@@ -1,7 +1,7 @@
 /** biome-ignore-all lint/performance/noNamespaceImport: Convenience */
 
 import { version } from "#package.json";
-import type { SessionSaveData, SystemSaveData } from "#system/game-data";
+import type { SessionSaveData, SystemSaveData } from "#types/data-types";
 import type { SessionSaveMigrator } from "#types/session-save-migrator";
 import type { SettingsSaveMigrator } from "#types/settings-save-migrator";
 import type { SystemSaveMigrator } from "#types/system-save-migrator";
