+++ src/system/version-migration/versions/v1_0_4.ts
@@ -3,8 +3,8 @@ import { allSpecies } from "#data/data-lists";
 import { CustomPokemonData } from "#data/pokemon-data";
 import { AbilityAttr } from "#enums/ability-attr";
 import { DexAttr } from "#enums/dex-attr";
-import type { SessionSaveData, SystemSaveData } from "#system/game-data";
 import { SettingKeys } from "#system/settings";
+import type { SessionSaveData, SystemSaveData } from "#types/data-types";
 import type { SessionSaveMigrator } from "#types/session-save-migrator";
 import type { SettingsSaveMigrator } from "#types/settings-save-migrator";
 import type { SystemSaveMigrator } from "#types/system-save-migrator";
