+++ src/system/version-migration/versions/v1_10_0.ts
@@ -2,7 +2,7 @@ import type { BattlerIndex } from "#enums/battler-index";
 import type { MoveId } from "#enums/move-id";
 import type { MoveResult } from "#enums/move-result";
 import { MoveUseMode } from "#enums/move-use-mode";
-import type { SessionSaveData } from "#system/game-data";
+import type { SessionSaveData } from "#types/data-types";
 import type { SessionSaveMigrator } from "#types/session-save-migrator";
 import type { TurnMove } from "#types/turn-move";
