+++ src/system/version-migration/versions/v1_9_0.ts
@@ -1,7 +1,7 @@
 import { MoveId } from "#enums/move-id";
 import { PokemonMove } from "#moves/pokemon-move";
-import type { SessionSaveData } from "#system/game-data";
 import type { PokemonData } from "#system/pokemon-data";
+import type { SessionSaveData } from "#types/data-types";
 import type { SessionSaveMigrator } from "#types/session-save-migrator";
 
 /**
