+++ src/ui/achvs-ui-handler.ts
@@ -5,9 +5,9 @@ import { TextStyle } from "#enums/text-style";
 import type { UiMode } from "#enums/ui-mode";
 import type { Achv } from "#system/achv";
 import { achvs, getAchievementDescription } from "#system/achv";
-import type { AchvUnlocks, VoucherUnlocks } from "#system/game-data";
 import type { Voucher } from "#system/voucher";
 import { getVoucherTypeIcon, getVoucherTypeName, vouchers } from "#system/voucher";
+import type { AchvUnlocks, VoucherUnlocks } from "#types/data-types";
 import { MessageUiHandler } from "#ui/message-ui-handler";
 import { ScrollBar } from "#ui/scroll-bar";
 import { addTextObject } from "#ui/text";
