+++ src/ui/pokedex-page-ui-handler.ts
@@ -43,8 +43,8 @@ import { TimeOfDay } from "#enums/time-of-day";
 import { UiMode } from "#enums/ui-mode";
 import type { Variant } from "#sprites/variant";
 import { getVariantIcon, getVariantTint } from "#sprites/variant";
-import type { StarterAttributes } from "#system/game-data";
 import { SettingKeyboard } from "#system/settings-keyboard";
+import type { StarterAttributes } from "#types/data-types";
 import type { DexEntry } from "#types/dex-data";
 import type { OptionSelectItem } from "#ui/abstract-option-select-ui-handler";
 import { BaseStatsOverlay } from "#ui/base-stats-overlay";
