+++ src/ui/pokedex-ui-handler.ts
@@ -31,8 +31,8 @@ import { UiMode } from "#enums/ui-mode";
 import { UiTheme } from "#enums/ui-theme";
 import type { Variant } from "#sprites/variant";
 import { getVariantIcon, getVariantTint } from "#sprites/variant";
-import type { DexAttrProps, StarterAttributes } from "#system/game-data";
 import { SettingKeyboard } from "#system/settings-keyboard";
+import type { DexAttrProps, StarterAttributes } from "#types/data-types";
 import type { DexEntry } from "#types/dex-data";
 import type { OptionSelectConfig } from "#ui/abstract-option-select-ui-handler";
 import { DropDown, DropDownLabel, DropDownOption, DropDownState, DropDownType, SortCriteria } from "#ui/dropdown";
