+++ src/ui/run-history-ui-handler.ts
@@ -6,8 +6,8 @@ import { PlayerGender } from "#enums/player-gender";
 import { TextStyle } from "#enums/text-style";
 import { TrainerVariant } from "#enums/trainer-variant";
 import { UiMode } from "#enums/ui-mode";
-import type { RunEntry } from "#system/game-data";
 import type { PokemonData } from "#system/pokemon-data";
+import type { RunEntry } from "#types/data-types";
 import { MessageUiHandler } from "#ui/message-ui-handler";
 import { RunDisplayMode } from "#ui/run-info-ui-handler";
 import { addTextObject } from "#ui/text";
