+++ test/plugins/api/pokerogue-session-savedata-api.test.ts
@@ -1,5 +1,4 @@
 import { PokerogueSessionSavedataApi } from "#api/pokerogue-session-savedata-api";
-import type { SessionSaveData } from "#system/game-data";
 import { initServerForApiTests } from "#test/test-utils/test-file-initialization";
 import { getApiBaseUrl } from "#test/test-utils/test-utils";
 import type {
@@ -10,6 +9,7 @@ import type {
   NewClearSessionSavedataRequest,
   UpdateSessionSavedataRequest,
 } from "#types/api/pokerogue-session-save-data-api";
+import type { SessionSaveData } from "#types/data-types";
 import { HttpResponse, http } from "msw";
 import type { SetupServerApi } from "msw/node";
 import { afterEach, beforeAll, beforeEach, describe, expect, it, vi } from "vitest";
