+++ test/test-utils/game-manager-utils.ts
@@ -8,7 +8,7 @@ import { GameModes } from "#enums/game-modes";
 import type { MoveId } from "#enums/move-id";
 import type { SpeciesId } from "#enums/species-id";
 import { PlayerPokemon } from "#field/pokemon";
-import type { StarterMoveset } from "#system/game-data";
+import type { StarterMoveset } from "#types/data-types";
 import type { Starter } from "#ui/starter-select-ui-handler";
 import { getPokemonSpecies, getPokemonSpeciesForm } from "#utils/pokemon-utils";
