+++ test/ui/pokedex.test.ts
@@ -6,8 +6,8 @@ import { DropDownColumn } from "#enums/drop-down-column";
 import { PokemonType } from "#enums/pokemon-type";
 import { SpeciesId } from "#enums/species-id";
 import { UiMode } from "#enums/ui-mode";
-import type { StarterAttributes } from "#system/game-data";
 import { GameManager } from "#test/test-utils/game-manager";
+import type { StarterAttributes } from "#types/data-types";
 import { FilterTextRow } from "#ui/filter-text";
 import { PokedexPageUiHandler } from "#ui/pokedex-page-ui-handler";
 import { PokedexUiHandler } from "#ui/pokedex-ui-handler";
